requirements/base.txt
# Bleeding edge Django
Mezzanine==4.0.1
-e git://github.com/thornomad/django-hitcount.git#egg=django-hitcount
django-simple-captcha
pyflakes
pep8
django_compressor
django-social-auth
django-jfu
python-magic
